Use lowercase for the type of target in publication artifactId
Not doing so was causing the publication to be unusable as for
some reason, the Gradle metadata in the .module files would
be all lowercase and resolved to non-existing URLs are they
would have uppercase characters as in iosSimulatorArm64.
Having those published with lowercase artifact ids should
address the issue.